+++ src/components/Alert/Alert.tsx
@@ -1,6 +1,6 @@
 //prettier-ignore
 /* Package imports */
-import React, { ReactElement, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 
 /* CSS imports */
 import "./Alert.scss";
@@ -10,13 +10,11 @@ import Button from "../Button";
 
 /* Assets imports*/
 import {
-  SvgCheck,
-  SvgError,
+  SvgIcConfirm,
+  SvgIcError,
   SvgIcErrorColored,
   SvgIcInfo,
-  SvgIcStatusLoading,
 } from "./../../assets/svg-components";
-import LoadingImage from "./../../assets/loader.gif";
 
 /* Interface */
 export interface AlertProps {
@@ -71,8 +69,8 @@ const Alert = (props: AlertProps) => {
      */
     const iconTypes = {
         info: SvgIcInfo,
-        success: SvgCheck,
-        warn: SvgError,
+        success: SvgIcConfirm,
+        warn: SvgIcError,
         error: SvgIcErrorColored
     };
 
@@ -194,7 +192,7 @@ const Alert = (props: AlertProps) => {
                 <div className={`n-alert-text-icon-wrapper`}>
                     {
                         loader ?
-                        <img src={LoadingImage} className={`n-alert-icon n-alert-loader`} /> :
+                        <img src={'https://cdn.pixelbin.io/v2/falling-surf-7c8bb8/fyprod/wrkr/misc/general/free/original/loader.gif'} className={`n-alert-icon n-alert-loader`} /> :
                         (
                             Icon || 
                             (